+++ workers/ohsome_quality_analyst/base/report.py
@@ -1,9 +1,11 @@
 import logging
+import os
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
 from statistics import mean
 from typing import Dict, List, Literal, NamedTuple, Tuple
 
+import jinja2
 from dacite import from_dict
 from geojson import Feature
 
@@ -28,6 +30,7 @@ class Result:
     label: Literal["green", "yellow", "red", "undefined"]
     value: float
     description: str
+    html: str
 
 
 class IndicatorLayer(NamedTuple):
@@ -48,7 +51,7 @@ class BaseReport(metaclass=ABCMeta):
         metadata = get_metadata("reports", type(self).__name__)
         self.metadata: Metadata = from_dict(data_class=Metadata, data=metadata)
         # Results will be written during the lifecycle of the report object (combine())
-        self.result = Result(None, None, None)
+        self.result = Result(None, None, None, None)
 
     def as_feature(self) -> Feature:
         """Returns a GeoJSON Feature object.
@@ -82,10 +85,13 @@ class BaseReport(metaclass=ABCMeta):
         logging.info(f"Combine indicators for report: {self.metadata.name}")
 
         values = []
+        html = ""
         for indicator in self.indicators:
             if indicator.result.label != "undefined":
                 values.append(indicator.result.value)
-
+            if indicator.result.html is not None:
+                html += indicator.result.html
+                del indicator.result.html
         if not values:
             self.result.value = None
             self.result.label = "undefined"
@@ -104,6 +110,24 @@ class BaseReport(metaclass=ABCMeta):
             self.result.label = "green"
             self.result.description = self.metadata.label_description["green"]
 
+        traffic_light, color = self.get_traffic_light_element()
+        template_dir = os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            "templates",
+        )
+        env = jinja2.Environment(
+            loader=jinja2.FileSystemLoader(template_dir),
+        )
+
+        template = env.get_template("report_template.html")
+        self.result.html = template.render(
+            report_name=self.metadata.name,
+            indicators=html,
+            result_description=self.result.description,
+            metadata=self.metadata.description,
+            traffic_light=traffic_light,
+        )
+
     @abstractmethod
     def set_indicator_layer(self) -> None:
         """Set the attribute indicator_layer."""
@@ -119,3 +143,50 @@ class BaseReport(metaclass=ABCMeta):
         attribution is necessary.
         """
         return get_attribution(["OSM"])
+
+    def get_traffic_light_element(self):
+        dot_css = (
+            "style='height: 25px; width: 25px; background-color: {0};"
+            "border-radius: 50%; display: inline-block;'"
+        )
+        traffic_light = (
+            "<span {0} class='dot'></span><span {1} class='dot'>"
+            "</span><span {2} class='dot'></span> Undefined Quality".format(
+                dot_css.format("#bbb"),
+                dot_css.format("#bbb"),
+                dot_css.format("#bbb"),
+            )
+        )
+        color = "grey"
+        if self.result.label == "red":
+            traffic_light = (
+                "<span {0} class='dot'></span><span {1} class='dot'>"
+                "</span><span {2} class='dot-red'></span> Bad Quality".format(
+                    dot_css.format("#FF0000"),
+                    dot_css.format("#bbb"),
+                    dot_css.format("#bbb"),
+                )
+            )
+            color = "red"
+        elif self.result.label == "yellow":
+            traffic_light = (
+                "<span {0} class='dot'></span><span {1} class='dot-yellow'>"
+                "</span><span {2} class='dot'></span> Medium Quality".format(
+                    dot_css.format("#bbb"),
+                    dot_css.format("#FFFF00"),
+                    dot_css.format("#bbb"),
+                )
+            )
+            color = "yellow"
+        elif self.result.label == "green":
+            traffic_light = (
+                "<span {0} class='dot-green'></span><span {1} class='dot'>"
+                "</span><span {2} class='dot'></span> Good Quality".format(
+                    dot_css.format("#008000"),
+                    dot_css.format("#bbb"),
+                    dot_css.format("#bbb"),
+                )
+            )
+            color = "green"
+
+        return traffic_light, color