+++ workers/tests/integrationtests/test_api_indicator.py
@@ -186,6 +186,23 @@ class TestApiIndicator(unittest.TestCase):
             content = response.json()
             self.assertEqual(content["type"], "RequestValidationError")
 
+    @oqt_vcr.use_cassette()
+    def test_indicator_return_as_html(self):
+        url = (
+            "/indicator?name={0}&layerName={1}&dataset={2}"
+            "&featureId={3}&fidField={4}&includeHtml={5}".format(
+                self.indicator_name,
+                self.layer_name,
+                self.dataset,
+                self.feature_id,
+                self.fid_field,
+                True,
+            )
+        )
+        response = self.client.get(url)
+        result = response.json()
+        self.assertIn("result.html", list(result["properties"].keys()))
+
 
 if __name__ == "__main__":
     unittest.main()