+++ workers/tests/integrationtests/test_api_report.py
@@ -25,7 +25,7 @@ class TestApiReport(unittest.TestCase):
         self.report_name = "SimpleReport"
         # Heidelberg
         self.dataset = "regions"
-        self.feature_id = "3"
+        self.feature_id = "9"
         self.fid_field = "ogc_fid"
 
         self.number_of_indicators = 2
@@ -155,6 +155,22 @@ class TestApiReport(unittest.TestCase):
                 self.assertEqual(response.status_code, 422)
                 content = response.json()
                 self.assertEqual(content["type"], "RequestValidationError")
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_include_html(self):
+        url = (
+            "/report?name={0}&dataset={1}&featureId={2}&fidField={3}"
+            "&includeHtml={4}".format(
+                self.report_name,
+                self.dataset,
+                self.feature_id,
+                self.fid_field,
+                True,
+            )
+        )
+        response = self.client.get(url)
+        result = response.json()
+        self.assertIn("report.result.html", list(result["properties"].keys()))
 
 
 if __name__ == "__main__":