+++ workers/tests/integrationtests/test_api_report_geojson_io.py
@@ -99,7 +99,7 @@ class TestApiReportIo(unittest.TestCase):
 
     @oqt_vcr.use_cassette()
     def test_report_include_svg(self):
-        feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
+        feature = get_geojson_fixture("niger-kanan-bakache.geojson")
         parameters = {
             "name": self.report_name,
             "bpolys": feature,
@@ -125,6 +125,36 @@ class TestApiReportIo(unittest.TestCase):
         response = self.client.post(self.endpoint, json=parameters)
         result = response.json()
         self.assertNotIn("indicators.0.result.svg", list(result["properties"].keys()))
+
+    @oqt_vcr.use_cassette()
+    def test_report_include_html(self):
+        feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
+        parameters = {
+            "name": self.report_name,
+            "bpolys": feature,
+            "includeSvg": False,
+            "includeHtml": True,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
+        result = response.json()
+        self.assertIn("report.result.html", list(result["properties"].keys()))
+        parameters = {
+            "name": self.report_name,
+            "bpolys": feature,
+            "includeSvg": False,
+            "includeHtml": False,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
+        result = response.json()
+        self.assertNotIn("report.result.html", list(result["properties"].keys()))
+
+        parameters = {
+            "name": self.report_name,
+            "bpolys": feature,
+        }
+        response = self.client.post(self.endpoint, json=parameters)
+        result = response.json()
+        self.assertNotIn("report.result.html", list(result["properties"].keys()))
 
 
 if __name__ == "__main__":