+++ workers/tests/unittests/test_report_sketchmap_fitness.py
@@ -17,6 +17,7 @@ class TestReportSketchmapFitness(unittest.TestCase):
             indicator = Mock()
             indicator.result = Mock()
             indicator.result.value = 0.5
+            indicator.result.html = "foo"
             report.indicators.append(indicator)
 
         report.combine_indicators()